refactor(preview): split launcher and add neutral frame boundary
