+++ TotalCrossSDK/src/main/java/totalcross/LauncherFonts.java
@@ -1,0 +1,8 @@
+// Copyright (C) 2026 Amalgam Solucoes em TI Ltda
+//
+// SPDX-License-Identifier: LGPL-2.1-only
+
+package totalcross;
+
+abstract class LauncherFonts extends LauncherSettings {
+}
